Split main file into modules
