+++ src/offline.rs
@@ -1,0 +1,32 @@
+use chess::{ChessMove, Game};
+use crate::utils::ask_for_move;
+
+pub(crate) fn offline_game() {
+    let mut game = Game::new();
+
+    // while the game is still ongoing
+    while game.result().is_none() {
+        let current_board = game.current_position();
+        println!("Current FEN: {}", current_board);
+
+        println!("{:?} to move. Enter the SAN move (ex: Nf3):", game.side_to_move());
+
+        let next_move_str = ask_for_move();
+
+        // convert the SAN to a valid move
+        // will repeat the loop if the move is not valid
+        let next_move = match ChessMove::from_san(&current_board, &next_move_str) {
+            Ok(m) => m,
+            Err(e) => {
+                println!("Problem with the chess move, try again. Error: {}", e);
+                continue;
+            },
+        };
+
+        // make the move
+        game.make_move(next_move);
+
+    }
+
+    println!("The game is over. Complete PGN for analysis: {}", game);
+}
