+++ src/online/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod game_setup;
+pub(crate) mod gameplay;
+pub(crate) mod commands;
